refactor(): move journeys out of DF dir
