+++ packages/forms-web-app/__tests__/unit/controllers/full-appeal/submit-appeal/list-of-documents.test.js
@@ -12,10 +12,7 @@ const {
 const { isFeatureActive } = require('../../../../../src/featureFlag');
 const { getDepartmentFromId } = require('../../../../../src/services/department.service');
 const { deleteAppeal } = require('../../../../../src/lib/appeals-api-wrapper');
-const {
-	baseS78SubmissionUrl,
-	taskListUrl
-} = require('../../../../../src/dynamic-forms/s78-appeal-form/journey');
+const { baseS78SubmissionUrl, taskListUrl } = require('../../../../../src/journs/s78-appeal-form');
 const { APPEALS_CASE_DATA } = require('@pins/common/src/constants');
 
 const apiClient = {
