+++ packages/forms-web-app/src/journs/has/appellant/index.js
@@ -1,10 +1,10 @@
-const { questions } = require('../questions');
-const { Journey } = require('../journey');
-const { Section } = require('../section');
+const { questions } = require('../../../dynamic-forms/questions');
+const { Journey } = require('../../../dynamic-forms/journey');
+const { Section } = require('../../../dynamic-forms/section');
 const {
 	questionHasAnswerBuilder,
 	questionsHaveAnswersBuilder
-} = require('../dynamic-components/utils/question-has-answer');
+} = require('../../../dynamic-forms/dynamic-components/utils/question-has-answer');
 
 const baseHASSubmissionUrl = '/appeals/householder';
 const taskListUrl = 'appeal-form/your-appeal';
@@ -14,7 +14,7 @@ const informationPageViewPath = 'dynamic-components/submission-information/index
 const journeyTitle = 'Appeal a planning decision';
 
 /**
- * @typedef {import('../journey-response').JourneyResponse} JourneyResponse
+ * @typedef {import('../../../dynamic-forms/journey-response').JourneyResponse} JourneyResponse
  */
 
 /**
