+++ packages/forms-web-app/src/journs/s78/lpa/index.js
@@ -1,10 +1,10 @@
-const { questions } = require('../questions');
-const { Journey } = require('../journey');
-const { Section } = require('../section');
+const { questions } = require('../../../dynamic-forms/questions');
+const { Journey } = require('../../../dynamic-forms/journey');
+const { Section } = require('../../../dynamic-forms/section');
 const {
 	questionHasAnswerBuilder,
 	questionsHaveAnswersBuilder
-} = require('../dynamic-components/utils/question-has-answer');
+} = require('../../../dynamic-forms/dynamic-components/utils/question-has-answer');
 
 const baseS78Url = '/manage-appeals/questionnaire';
 const s78JourneyTemplate = 'questionnaire-template.njk';
@@ -14,7 +14,7 @@ const journeyTitle = 'Manage your appeals';
 const { APPEAL_CASE_PROCEDURE } = require('pins-data-model');
 
 /**
- * @typedef {import('../journey-response').JourneyResponse} JourneyResponse
+ * @typedef {import('../../../dynamic-forms/journey-response').JourneyResponse} JourneyResponse
  */
 
 /**
